Extract module jawanndenn.views.static_files
